requirements.txt
streamlit
langchain
langchain_google_genai
pypdf
beautifulsoup4
requests
lxml
python-dotenv